Bug fix and improvements
Fixed a bug when saving the sky backgrounds: the file name was the same as for internal background. The file names are now distinct.

Added a mask in the generation of flats to compensate the vignetting of the IFU. The values outside the mask are replaced by a value of one. The use of the mask can be controlled in the call to compute_detector_flat(). By default the mask is applied.